+++ CHANGELOG.rst
@@ -35,8 +35,9 @@ Motions:
    follow recommendation, manage submitters and supporters, change motion
    category, motion block and origin and manage motion polls [#3913].
  - Added new permission to create amendments [#4128].
- - Added new flag to motion state to control access for different users. Added
-   new permission to see motions in some internal state [#4235, #4518].
+ - Added new flag to motion state to control access restrictions for different
+   users. Added new permission to see motions in some internal state
+   [#4235, #4518, #4521].
  - Allowed submitters to set state of new motions in complex and customized
    workflow [#4236].
  - Added multi select action to manage submitters, tags, states and